api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
azure-storage-blob==12.19.0
pandas==2.1.4
pyarrow==14.0.1
python-dotenv==1.0.0